+++ src/main/java/com/example/movie_rating_service/model/Review.java
@@ -18,7 +18,7 @@ public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "review_id")
-    private int reviewId;
+    private long reviewId;
 
     @Id
     @ManyToOne
